requirements.txt
streamlit>=1.28.0
llama-parse>=0.4.0
google-generativeai>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0
